+++ applications/fishing-map/src/features/workspaces-list/WorkspacesList.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useCallback } from 'react'
 import cx from 'classnames'
 import { useSelector } from 'react-redux'
 import Link from 'redux-first-router-link'
@@ -7,18 +7,32 @@ import { Spinner } from '@globalfishingwatch/ui-components'
 import { isValidLocationCategory, selectLocationCategory } from 'routes/routes.selectors'
 import { HOME, WORKSPACE } from 'routes/routes'
 import { AsyncReducerStatus } from 'utils/async-slice'
-import { DEFAULT_WORKSPACE_KEY } from 'data/workspaces'
+import { DEFAULT_WORKSPACE_ID } from 'data/workspaces'
+import useViewport from 'features/map/map-viewport.hooks'
 import styles from './WorkspacesList.module.css'
-import { selectCurrentHighlightedWorkspaces } from './workspaces-list.selectors'
+import {
+  HighlightedWorkspaceMerged,
+  selectCurrentHighlightedWorkspaces,
+} from './workspaces-list.selectors'
 import { selectHighlightedWorkspacesStatus } from './workspaces-list.slice'
 
 function WorkspacesList() {
   const { t } = useTranslation()
+  const { setMapCoordinates } = useViewport()
   const locationCategory = useSelector(selectLocationCategory)
   const userFriendlyCategory = locationCategory.replace('-', ' ')
   const highlightedWorkspaces = useSelector(selectCurrentHighlightedWorkspaces)
   const highlightedWorkspacesStatus = useSelector(selectHighlightedWorkspacesStatus)
   const validCategory = useSelector(isValidLocationCategory)
+
+  const onWorkspaceClick = useCallback(
+    (workspace: HighlightedWorkspaceMerged) => {
+      if (workspace.viewport) {
+        setMapCoordinates(workspace.viewport)
+      }
+    },
+    [setMapCoordinates]
+  )
 
   if (highlightedWorkspacesStatus === AsyncReducerStatus.Finished && !validCategory) {
     return (
@@ -42,7 +56,7 @@ function WorkspacesList() {
           {highlightedWorkspaces?.map((highlightedWorkspace) => {
             const active = highlightedWorkspace?.id !== undefined
             const linkTo =
-              highlightedWorkspace.id === DEFAULT_WORKSPACE_KEY
+              highlightedWorkspace.id === DEFAULT_WORKSPACE_ID
                 ? {
                     type: HOME,
                     payload: {},
@@ -57,20 +71,28 @@ function WorkspacesList() {
                     query: {},
                   }
             return (
-              <li key={highlightedWorkspace.name}>
+              <li key={highlightedWorkspace.id || highlightedWorkspace.name}>
                 <div className={cx(styles.workspace, { [styles.disabled]: !active })}>
-                  <img
-                    className={styles.image}
-                    alt={highlightedWorkspace.name}
-                    src={highlightedWorkspace.img}
-                  />
+                  <Link to={linkTo} onClick={() => onWorkspaceClick(highlightedWorkspace)}>
+                    <img
+                      className={styles.image}
+                      alt={highlightedWorkspace.name}
+                      src={highlightedWorkspace.img}
+                    />
+                  </Link>
                   <div className={styles.info}>
-                    <h3 className={styles.title}>{highlightedWorkspace.name}</h3>
+                    <Link to={linkTo} onClick={() => onWorkspaceClick(highlightedWorkspace)}>
+                      <h3 className={styles.title}>{highlightedWorkspace.name}</h3>
+                    </Link>
                     <p
                       className={styles.description}
                       dangerouslySetInnerHTML={{ __html: highlightedWorkspace.description }}
                     ></p>
-                    <Link to={linkTo} className={styles.link}>
+                    <Link
+                      to={linkTo}
+                      className={styles.link}
+                      onClick={() => onWorkspaceClick(highlightedWorkspace)}
+                    >
                       {highlightedWorkspace.cta}
                     </Link>
                   </div>