+++ applications/fishing-map/src/utils/info.ts
@@ -22,7 +22,12 @@ export const formatNumber = (num: string | number) => {
   })
 }
 
-export const getVesselLabel = (vessel: ExtendedFeatureVessel | Vessel) => {
+export const getVesselLabel = (vessel: ExtendedFeatureVessel | Vessel, withGearType = false) => {
+  if (vessel.shipname && vessel.geartype && withGearType)
+    return `${formatInfoField(vessel.shipname, 'name')} (${i18n.t(
+      `vessel.gearTypes.${vessel.geartype}` as any,
+      '---'
+    )})`
   if (vessel.shipname) return formatInfoField(vessel.shipname, 'name')
   if (vessel.registeredGearType) {
     return `${i18n.t('vessel.unkwownVesselByGeartype', {