+++ packages/react-hooks/src/use-map-interaction/use-map-interaction.ts
@@ -8,6 +8,21 @@ import { ExtendedFeature, InteractionEventCallback, InteractionEvent } from '.'
 type FeatureStates = 'click' | 'hover' | 'highlight'
 type FeatureStateSource = { source: string; sourceLayer: string; id: string; state?: FeatureStates }
 
+export const filterUniqueFeatureInteraction = (features: ExtendedFeature[]) => {
+  const uniqueLayerIdFeatures: Record<string, boolean> = {}
+  const filtered = features?.filter(({ layerId, uniqueFeatureInteraction }) => {
+    if (!uniqueFeatureInteraction) {
+      return true
+    }
+    if (uniqueLayerIdFeatures[layerId] === undefined) {
+      uniqueLayerIdFeatures[layerId] = true
+      return true
+    }
+    return false
+  })
+  return filtered
+}
+
 const getExtendedFeatures = (
   features: MapboxGeoJSONFeature[],
   metadata?: ExtendedStyleMeta,
@@ -26,6 +41,8 @@ const getExtendedFeatures = (
     }
 
     const unit = generatorMetadata?.legend?.unit ?? null
+
+    const uniqueFeatureInteraction = feature.layer?.metadata?.uniqueFeatureInteraction ?? false
     const properties = feature.properties || {}
     const extendedFeature: ExtendedFeature | null = {
       properties,
@@ -34,6 +51,7 @@ const getExtendedFeatures = (
       layerId: feature.layer.id,
       source: feature.source,
       sourceLayer: feature.sourceLayer,
+      uniqueFeatureInteraction,
       id: (feature.id as number) || feature.properties?.gfw_id || undefined,
       value: properties.value || properties.name || properties.id,
       unit,
@@ -164,9 +182,11 @@ export const useMapClick = (
           metadata,
           true
         )
-        if (extendedFeatures.length) {
-          interactionEvent.features = extendedFeatures
-          updateFeatureState(extendedFeatures, 'click')
+        const extendedFeaturesLimit = filterUniqueFeatureInteraction(extendedFeatures)
+
+        if (extendedFeaturesLimit.length) {
+          interactionEvent.features = extendedFeaturesLimit
+          updateFeatureState(extendedFeaturesLimit, 'click')
         }
       }
       clickCallback(interactionEvent)
@@ -178,7 +198,7 @@ export const useMapClick = (
 }
 
 type MapHoverConfig = {
-  debounced: number
+  debounced?: number
 }
 export const useMapHover = (
   hoverCallbackImmediate?: InteractionEventCallback,
@@ -211,12 +231,13 @@ export const useMapHover = (
       }
       if (event.features?.length) {
         const extendedFeatures: ExtendedFeature[] = getExtendedFeatures(event.features, metadata)
+        const extendedFeaturesLimit = filterUniqueFeatureInteraction(extendedFeatures)
 
-        if (extendedFeatures.length) {
-          hoverEvent.features = extendedFeatures
+        if (extendedFeaturesLimit.length) {
+          hoverEvent.features = extendedFeaturesLimit
         }
 
-        updateFeatureState(extendedFeatures, 'hover')
+        updateFeatureState(extendedFeaturesLimit, 'hover')
       }
 
       if (hoverCallbackDebounced?.current) {